refactor(engine): split unit tests into domain modules under 1k
Decompose the 1554-line engine test monolith into helpers plus focused
modules (settings, drawing validation, scoring, phases/reconnect,
actions, reactions). Extract shared reach/play helpers and tighten
weak assertions from the adversarial sweep.
