+++ server/src/engine/tests/mod.rs
@@ -1,0 +1,10 @@
+//! Engine unit tests — split by domain so the suite can grow without a monolith.
+
+mod helpers;
+
+mod actions;
+mod drawing_validation;
+mod phases_reconnect;
+mod reactions;
+mod scoring;
+mod settings;
